+++ languagetool-standalone/CHANGES.md
@@ -13,7 +13,12 @@
   * The unmaintained code from package `org.languagetool.dev.wikipedia.atom`
     has been removed. It hadn't been maintained for years and didn't work properly
     anymore.
-  
+
+#### Java API
+  * `AbstractSimpleReplaceRule2` has been fixed so that it's now case-insensitive.
+    If you implement a sub class of it and you want the old behavior, please implement
+    `isCaseSensitive()` and have it return `true`. (Issue #2051)
+    
 #### Internal
   * Experimental: the new `default="temp_off"` attribute in `grammar.xml` files will
     turn off a rule/rulegroup, but keep it activated for our nightly regression tests.